requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
faiss-cpu==1.7.4
filelock==3.18.0
fsspec==2024.12.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json_repair==0.39.1
loguru==0.7.2
MarkupSafe==3.0.2
mpmath==1.3.0
neo4j==5.20.0
neo4j-graphrag==1.8.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openai==1.35.3
packaging==25.0
pillow==11.3.0
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==5.8.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.53.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
urllib3==2.5.0
